requirements.txt
flask
gunicorn
python-dotenv
pydantic
openai
anthropic
langchain
langchain-openai
langchain-anthropic
streamlit
